Html and javascript changes for tectboxt and event listeners
I have created new Ids for the my text area so that I can be used a s a n event listener  and the input boxes that could can be replaced by labels later and in the Javascript I removed redundant code that would not help me such as the choices being case insensitve because we are not relying on human input that needs to be typed instead we need it to be dependant on what button is pressed.
I also changed some of the choices from being just rock , paper or scissors and made them all uppercase to keep everything simple and consistant. I moved the event listeners to the palygame function so that when the button is clicked it will indicate what was seleceted by the user and to do in response and where. The event listener that would display hello is a test to see if it would work properly but initally it did not display in the text area however after some research if I included my other event listeners in a new event listen that uses DOMContentLoaded would allow my message to display in the text area.